utilsファイルの残りの型移動
