+++ common/types/ui-elements/image.ts
@@ -1,0 +1,9 @@
+import { StaticImageData } from 'next/image';
+
+export type ImageWrapperProps = {
+  src: string | StaticImageData;
+  alt: string;
+  width: number;
+  height: number;
+  className?: string;
+};
