+++ context/FormContext.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useState, useMemo } from 'react';
-import { ChildrenProps } from '@/common/types/utils';
 import { FormType } from '@/common/types/ui-parts/form';
 import { FormContextProps } from '@/common/types/context';
+import { ChildrenProps } from '@/common/types/ui-parts/layout';
 
 const InitialState: FormContextProps = {
   formType: 'createStack',
