+++ features/introspections/contexts/StackIntrospectionContext.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useState, useMemo } from 'react';
-import { ChildrenProps } from '@/common/types/utils';
 import { StackIntrospectionContextProps } from '../types/context';
 import { IntrospectionProps } from '../types/introspection';
+import { ChildrenProps } from '@/common/types/ui-parts/layout';
 
 const InitialState: StackIntrospectionContextProps = {
   showStackIntrospection: undefined,
