+++ features/outputs/contexts/OutputFormContext.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useState, useMemo } from 'react';
-import { ChildrenProps } from '@/common/types/utils';
 import { InitialOutputFormData } from '../functions/form';
 import { OutputFormContextProps, OutputFormDataParams } from '../types/context';
+import { ChildrenProps } from '@/common/types/ui-parts/layout';
 
 const InitialState: OutputFormContextProps = {
   outputFormData: InitialOutputFormData,
